% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiModelHIV-package.R
\docType{package}
\name{EpiModelHIV.campcl-package}
\alias{EpiModelHIV.campcl-package}
\alias{EpiModelHIV.campcl}
\title{HIV Transmission Dynamics among adult and adolecent MSM}
\description{
\tabular{ll}{
   Package: \tab EpiModelHIV.campcl\cr
   Type: \tab Package\cr
   Version: \tab 1.0.0\cr
   Date: \tab 2016-06-25\cr
   License: \tab GPL (>= 2)\cr
   LazyLoad: \tab yes\cr
}
}
\details{
EpiModelHIV.campcl provides a version of the EpiModelHIV package for
for simulating HIV transmission dynamics among adult and adolecent men who have sex with men (MSM) 
in the United States.
}
\keyword{asmm}
\keyword{msm}
\keyword{package}
